refactor(capabilities): :recycle: migrate accordion, animation, breadcrumbs molecules to component-store pattern
- accordion: ComponentStore<{isOpen,panelId,triggerId},{toggle}> via useAccordionStore
- animation/motion-div: nested under category dir, StateStore<{motionEnabled}> via useMotionDivStore
- breadcrumbs/breadcrumb: nested under category dir, StateStore<{navRef,isVisible}> via useBreadcrumbStore; useGlassmorphism allowlisted in component
- Add ERROR-level ESLint rules for molecules tier (jsx-no-bind, prefer-component-store, folder-composition, store-return-shape, index-only-component)
- Add ERROR-level depcruise rules for molecules tier (import-only-via-index, seal-private-nested-folders, design-system-layering-molecules)
- Fix organism/reading-content-progress-bar to import terminal-progress-bar via index
- Fix videogames/content.mdx to import matrix-terminal without .tsx extension
