+++ src/components/design-system/molecules/accordion/accordion/index.ts
@@ -1,0 +1,1 @@
+export { Accordion } from "./accordion";
