+++ src/components/design-system/molecules/animation/motion-div/index.ts
@@ -1,0 +1,1 @@
+export { MotionDiv } from "./motion-div";
