+++ src/components/design-system/molecules/animation/motion-div/motion-div.tsx
@@ -2,12 +2,13 @@
 
 import { motion, HTMLMotionProps } from "framer-motion";
 import { ComponentPropsWithoutRef, FC } from "react";
-import { useMotionStore } from "@/components/design-system/hooks/use-motion-store";
+import { useMotionDivStore } from "./use-motion-div-store";
 
 type MotionDivProps = HTMLMotionProps<"div">;
 
 export const MotionDiv: FC<MotionDivProps> = (props) => {
-    const motionEnabled = useMotionStore();
+    const { state } = useMotionDivStore();
+    const { motionEnabled } = state;
 
     if (!motionEnabled) {
         /* eslint-disable @typescript-eslint/no-unused-vars */
