+++ src/components/design-system/molecules/animation/motion-div/use-motion-div-store.ts
@@ -1,0 +1,13 @@
+"use client";
+
+import { useMotionStore } from "@/components/design-system/hooks/use-motion-store";
+import { StateStore } from "@/types/component-store";
+
+interface MotionDivState {
+    motionEnabled: boolean;
+}
+
+export const useMotionDivStore = (): StateStore<MotionDivState> => {
+    const motionEnabled = useMotionStore();
+    return { state: { motionEnabled } };
+};
