+++ src/components/design-system/molecules/breadcrumbs/breadcrumb/index.ts
@@ -1,0 +1,2 @@
+export { Breadcrumb } from "./breadcrumb";
+export type { BreadcrumbItem } from "./breadcrumb";
