+++ src/content/videogames/content.mdx
@@ -7,7 +7,7 @@ tags: [two pointers, data structures, algorithms]
 authors: [fabrizio_duroni]
 ---
 
-import { MatrixTerminal } from '@/components/design-system/molecules/effects/matrix-terminal.tsx'
+import { MatrixTerminal } from '@/components/design-system/molecules/effects/matrix-terminal'
 
 # Videogames
